Migrated build.gradle from Groovy to the Kotlin
Added Border Activity

+++ app/src/test/java/com/example/first_jetcompose/ExampleUnitTest.kt
@@ -1,8 +1,7 @@
 package com.example.first_jetcompose
 
 import org.junit.Test
-
-import org.junit.Assert.*
+import org.junit.Assert.assertEquals
 
 /**
  * Example local unit test, which will execute on the development machine (host).
@@ -12,6 +11,6 @@ import org.junit.Assert.*
 class ExampleUnitTest {
     @Test
     fun addition_isCorrect() {
-        assertEquals(4, 2 + 2)
+        assertEquals(4, (2 + 2).toInt())
     }
 }
